xcp.repository: switch from ConfigParser to configparser (has backport)
This commit was originally authored by Yann Dirson (see below).

This is supposed to be just a module renaming to conform to PEP8, see
https://docs.python.org/3/whatsnew/3.0.html#library-changes

The SafeConfigParser class has been renamed to ConfigParser in Python
3.2, and backported as addon package.  The `readfp` method now
triggers a deprecation warning to replace it with `read_file`.

This commit supports accessing the repository text files using Python3
as well as Python2.

Note: Reading of binary files will be fixed in the subsequent commit.

Co-authored-by: Yann Dirson <yann.dirson@vates.fr>
Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>